SoundEx code
This code is based on the algorithm SoundEx. SoundEx is a phonetic algorithm for indexing names by sound, as pronounced in English.

Some of the American Soundex principles was used to better understand and developer this algorithm. Such as the GenerateGroupKey method, where each group of letters is replaced by a number. This method helps to compare and define each word is phonetic similar.